requirements.txt
# Web Framework
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0

# Database
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
SQLAlchemy==1.4.23

# Environment
python-dotenv==0.19.0

# File processing
python-docx==0.8.11
fpdf2==2.5.1
PyPDF2==3.0.1
nltk==3.6.3

# AI & ML
openai==0.27.0
scikit-learn==1.0.0
numpy==1.21.2

# Utilities
Werkzeug==2.0.1
requests==2.26.0
pandas==1.3.3
